+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.1] - 2024-08-05
+
+### Changed
+
+- Fix `shouldTrackEvents` method. Now the `defaultPayload` and `Payload` are merged before sending to the validation `shouldTrack`.
+
 ## [1.1.0] - 2024-08-02
 
 ### Added